add: database crate added

+++ crates/database/src/entities/twine_batch_l2_blocks.rs
@@ -3,9 +3,9 @@
 use sea_orm::entity::prelude::*;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
-#[sea_orm(table_name = "twine_batch_l2_transactions")]
+#[sea_orm(table_name = "twine_batch_l2_blocks")]
 pub struct Model {
-    pub batch_number: i32,
+    pub batch_number: i64,
     #[sea_orm(
         primary_key,
         auto_increment = false,
